add settings toggle to not delete imported archives

+++ osu.Game/Database/RealmArchiveModelImporter.cs
@@ -8,11 +8,14 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Humanizer;
+using osu.Framework.Allocation;
+using osu.Framework.Bindables;
 using osu.Framework.Extensions;
 using osu.Framework.Extensions.IEnumerableExtensions;
 using osu.Framework.Logging;
 using osu.Framework.Platform;
 using osu.Framework.Threading;
+using osu.Game.Configuration;
 using osu.Game.Extensions;
 using osu.Game.IO.Archives;
 using osu.Game.Models;
@@ -77,11 +80,19 @@ namespace osu.Game.Database
         /// </summary>
         public Action<Notification>? PostNotification { get; set; }
 
-        protected RealmArchiveModelImporter(Storage storage, RealmAccess realm)
+        private readonly OsuConfigManager? config;
+
+        private Bindable<bool>? deleteImportedArchives;
+
+        protected RealmArchiveModelImporter(Storage storage, RealmAccess realm, OsuConfigManager? config = null)
         {
             Realm = realm;
 
             Files = new RealmFileStore(realm, storage);
+
+            deleteImportedArchives = config?.GetBindable<bool>(OsuSetting.DeleteImportedArchives);
+
+            this.config = config;
         }
 
         public Task Import(params string[] paths) => Import(paths.Select(p => new ImportTask(p)).ToArray());
@@ -241,9 +252,10 @@ namespace osu.Game.Database
             //  e.g. reconstructing/repairing database with items from default storage.
             // Also, not always a single file, i.e. for LegacyFilesystemReader
             // TODO: Add a check to prevent files from storage to be deleted.
+            bool allowDelete = deleteImportedArchives?.Value ?? true;
             try
             {
-                if (import != null && ShouldDeleteArchive(task.Path))
+                if (import != null && ShouldDeleteArchive(task.Path) && allowDelete)
                     task.DeleteFile();
             }
             catch (Exception e)
